moved the API calls to services folder
