+++ src/Services/Vertex.tsx
@@ -14,36 +14,15 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-import { toast } from 'react-toastify';
-import { JupyterLab } from '@jupyterlab/application';
-import { requestAPI } from '../../handler/handler';
-import { DataprocLoggingService, LOG_LEVEL } from '../../utils/loggingService';
-import { toastifyCustomStyle } from '../../utils/utils';
-import { Dayjs } from 'dayjs';
-import { ICreatePayload, IDagList, IDagRunList, IDeleteSchedulerAPIResponse, IMachineType, ITriggerSchedule, IUpdateSchedulerAPIResponse } from './VertexInterfaces';
+import { JupyterLab } from "@jupyterlab/application";
+import { toast } from "react-toastify";
+import { requestAPI } from "../handler/handler";
+import { ICreatePayload, IDagList, IDagRunList, IDeleteSchedulerAPIResponse, IMachineType, ITriggerSchedule, IUpdateSchedulerAPIResponse } from "../scheduler/VertexScheduler/VertexInterfaces";
+import { DataprocLoggingService, LOG_LEVEL } from "../utils/loggingService";
+import { toastifyCustomStyle } from "../utils/utils";
+import { Dayjs } from "dayjs";
 
 export class VertexServices {
-    static getParentProjectAPIService = async (
-        setHostProject: (value: string) => void,
-    ) => {
-        try {
-            const formattedResponse: any = await requestAPI(`api/compute/getXpnHost`);
-            if (formattedResponse.length === 0) {
-                setHostProject('')
-            } else {
-                if (formattedResponse) {
-                    setHostProject(formattedResponse);
-                }
-            }
-        } catch (error) {
-            DataprocLoggingService.log(
-                'Error fetching host project',
-                LOG_LEVEL.ERROR
-            );
-            setHostProject('')
-            toast.error(`Failed to fetch host project`);
-        }
-    };
     static machineTypeAPIService = async (
         region: string,
         setMachineTypeList: (value: IMachineType[]) => void,
@@ -73,167 +52,6 @@ export class VertexServices {
             );
         }
     };
-
-    static cloudStorageAPIService = async (
-        setCloudStorageList: (value: string[]) => void,
-        setCloudStorageLoading: (value: boolean) => void
-    ) => {
-        try {
-            setCloudStorageLoading(true)
-            const formattedResponse: any = await requestAPI(`api/storage/listBucket`);
-            if (formattedResponse.length > 0) {
-                setCloudStorageList(formattedResponse)
-            } else {
-                setCloudStorageList([])
-            }
-            setCloudStorageLoading(false)
-        } catch (error) {
-            setCloudStorageList([])
-            setCloudStorageLoading(false)
-            DataprocLoggingService.log(
-                'Error listing cloud storage bucket',
-                LOG_LEVEL.ERROR
-            );
-            toast.error(
-                `Failed to fetch cloud storage bucket`,
-                toastifyCustomStyle
-            );
-        }
-    };
-
-    static serviceAccountAPIService = async (
-        setServiceAccountList: (
-            value: { displayName: string; email: string }[]
-        ) => void,
-        setServiceAccountLoading: (value: boolean) => void
-    ) => {
-        try {
-            setServiceAccountLoading(true)
-            const formattedResponse: any = await requestAPI(`api/iam/listServiceAccount`);
-            if (formattedResponse.length === 0) {
-                setServiceAccountList([])
-            } else {
-                const serviceAccountList = formattedResponse.map((account: any) => ({
-                    displayName: account.displayName,
-                    email: account.email
-                }));
-                serviceAccountList.sort();
-                setServiceAccountList(serviceAccountList);
-            }
-            setServiceAccountLoading(false)
-        } catch (error) {
-            setServiceAccountList([])
-            setServiceAccountLoading(false)
-            DataprocLoggingService.log(
-                'Error listing service accounts',
-                LOG_LEVEL.ERROR
-            );
-            toast.error(
-                `Failed to fetch service accounts list`,
-                toastifyCustomStyle
-            );
-        }
-    };
-
-    static primaryNetworkAPIService = async (
-        setPrimaryNetworkList: (value: { name: string; link: string }[]) => void,
-        setPrimaryNetworkLoading: (value: boolean) => void
-    ) => {
-        try {
-            setPrimaryNetworkLoading(true)
-            const formattedResponse: any = await requestAPI(`api/compute/network`);
-            if (formattedResponse.length === 0) {
-                setPrimaryNetworkList([])
-            } else {
-                const primaryNetworkList = formattedResponse.map((network: any) => ({
-                    name: network.name,
-                    link: network.selfLink
-                }));
-                primaryNetworkList.sort();
-                setPrimaryNetworkList(primaryNetworkList);
-            }
-            setPrimaryNetworkLoading(false)
-        } catch (error) {
-            setPrimaryNetworkList([])
-            setPrimaryNetworkLoading(false)
-            DataprocLoggingService.log(
-                'Error listing primary network',
-                LOG_LEVEL.ERROR
-            );
-            toast.error(
-                `Failed to fetch primary network list`,
-                toastifyCustomStyle
-            );
-        }
-    };
-
-    static subNetworkAPIService = async (
-        region: string,
-        primaryNetworkSelected: string | undefined,
-        setSubNetworkList: (value: { name: string; link: string }[]) => void,
-        setSubNetworkLoading: (value: boolean) => void
-    ) => {
-        try {
-            setSubNetworkLoading(true)
-            const formattedResponse: any = await requestAPI(`api/compute/subNetwork?region_id=${region}&network_id=${primaryNetworkSelected}`);
-            if (formattedResponse.length === 0) {
-                setSubNetworkList([])
-            } else {
-                const subNetworkList = formattedResponse.map((network: any) => ({
-                    name: network.name,
-                    link: network.selfLink
-                }));
-                subNetworkList.sort();
-                setSubNetworkList(subNetworkList);
-            }
-            setSubNetworkLoading(false)
-        } catch (error) {
-            setSubNetworkList([])
-            setSubNetworkLoading(false)
-            DataprocLoggingService.log(
-                'Error listing sub networks',
-                LOG_LEVEL.ERROR
-            );
-            toast.error(
-                `Failed to fetch sub networks list`,
-                toastifyCustomStyle
-            );
-        }
-    };
-
-    static sharedNetworkAPIService = async (
-        setSharedNetworkList: (value: { name: string; network: string, subnetwork: string }[]) => void,
-        setSharedNetworkLoading: (value: boolean) => void
-    ) => {
-        try {
-            setSharedNetworkLoading(true)
-            const formattedResponse: any = await requestAPI(`api/compute/sharedNetwork`);
-            if (formattedResponse.length === 0) {
-                setSharedNetworkList([])
-            } else {
-                const sharedNetworkList = formattedResponse.map((network: any) => ({
-                    name: network.network.split('/').pop(),
-                    network: network.network,
-                    subnetwork: network.subnetwork
-                }));
-                sharedNetworkList.sort();
-                setSharedNetworkList(sharedNetworkList);
-            }
-            setSharedNetworkLoading(false)
-        } catch (error) {
-            setSharedNetworkList([])
-            setSharedNetworkLoading(false)
-            DataprocLoggingService.log(
-                'Error listing shared networks',
-                LOG_LEVEL.ERROR
-            );
-            toast.error(
-                `Failed to fetch shared networks list`,
-                toastifyCustomStyle
-            );
-        }
-    };
-
     static createVertexSchedulerService = async (
         payload: ICreatePayload,
         app: JupyterLab,
@@ -273,7 +91,6 @@ export class VertexServices {
             );
         }
     };
-
     static listVertexSchedules = async (
         setDagList: (value: IDagList[]) => void,
         region: string,
@@ -302,7 +119,6 @@ export class VertexServices {
 
         }
     }
-
     static handleUpdateSchedulerPauseAPIService = async (
         scheduleId: string,
         region: string,
@@ -463,7 +279,6 @@ export class VertexServices {
             );
         }
     };
-
     static executionHistoryServiceList = async (
         region: string,
         schedulerData: any,
@@ -568,39 +383,5 @@ export class VertexServices {
             );
         }
         setIsLoading(false)
-    };
-
-    static vertexJobTaskLogsListService = async (
-        dagRunId: string | undefined,
-        jobRunsData: IDagRunList | undefined,
-        setDagTaskInstancesList: (value: any) => void,
-        setIsLoading: (value: boolean) => void
-    ) => {
-        setDagTaskInstancesList([]);
-        setIsLoading(true);
-        const start_date = encodeURIComponent(jobRunsData?.startDate || '');
-        const end_date = encodeURIComponent(jobRunsData?.endDate || '');
-        try {
-            const data: any = await requestAPI(
-                `api/logEntries/listEntries?filter_query=timestamp >= \"${start_date}" AND timestamp <= \"${end_date}" AND SEARCH(\"${dagRunId}\")`
-            );
-            let transformDagRunTaskInstanceListData = [];
-            transformDagRunTaskInstanceListData = data.map(
-                (dagRunTask: any) => {
-                    return {
-                        severity: dagRunTask.severity,
-                        textPayload: dagRunTask.textPayload && dagRunTask.textPayload ? dagRunTask.textPayload : '',
-                        date: new Date(dagRunTask.timestamp).toDateString(),
-                        time: new Date(dagRunTask.timestamp).toTimeString().split(' ')[0],
-                        fullData: dagRunTask,
-                    };
-                }
-            );
-            setDagTaskInstancesList(transformDagRunTaskInstanceListData);
-            setIsLoading(false);
-        } catch (reason) {
-            setIsLoading(false);
-            setDagTaskInstancesList([]);
-        }
     };
 }
